requirements.txt
alembic==1.14.1
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.1
dnspython==2.7.0
email_validator==2.2.0
Flask==3.1.0
Flask-Cors==5.0.0
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
Mako==1.3.9
MarkupSafe==3.0.2
oauthlib==3.2.2
pycparser==2.22
PyJWT==2.10.1
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
SQLAlchemy==2.0.38
typing_extensions==4.12.2
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
urllib3==2.3.0
user-agents==2.2.0
Werkzeug==3.1.3
WTForms==3.2.1
